+++ CHANGELOG.md
@@ -6,6 +6,9 @@ NOTE: Version needs to be updated in the following places:
 - [ ] Compatibility.version constant (must be hard coded since inaccessible in code)
 - [ ] Update changelog and tag with matching version in GitHub.
 
+v1.6.4 1/17/2025 Added debugging output when replacing the identifier in preview/playground environment to fix issue with Score identifier being com.kudit.Score-.  Added check to prevent preview output alerting that iCloud doesn't work from spamming the logs.  Added in app name and identifier to compatibility info.  Fixed unnecessary check for iOS warning in Backport.
+
+*PASSES ALL SWIFTPACKAGEINDEX TESTS*
 v1.6.3 1/15/2025 Added some documentation to `asJSON()` function.  Fixed internal definition of Triangle initializer.
 
 *PASSES ALL SWIFTPACKAGEINDEX TESTS*